plate type checkpoint
